refactor: Use go 1.21 and new slices.Contains func
This also includes showing how to use different flake versions, necessary
in this case because golangci-lint needed to be built with go 1.21 to
support go 1.21, but the available prebuilt nix package was still being
built with go 1.20. This can be removed later, but is a good experiment
to show how flexible devbox is.