feat: vendor infrastructure provider APIs (#190)
@jimmidyson and I were briefly talking about this idea. Here is an
attempt of using infra provider types that are synced from upstream
repos. It is easier to work with Go types instead of `unstructured`.
Switched the AWS region handler to use these new types to show that it
works.

An important part of this PR is the automated syncing. I added it as
makefile targets to keep it simple, but can be extended if no longer
enough as we add additional providers.

---------

Co-authored-by: Jimmi Dyson <jimmidyson@gmail.com>