fix: Remove vendored types for core CAPI providers (CAPD, CABPK, KCP) (#452)
We do not need to sync the Kubeadm Control Plane or Kubeadm Bootstrap
provider types, because they are developed in the sigs.k8s.io/cluster-api module, which
we define as a dependency.

We also do not need to sync the Docker infrastructure provider types,
because they in the sigs.k8s.io/cluster-api/test module, which is developed
in parallel with sigs.k8s.io/cluster-api. For that reason, we expect no dependency
conflicts.