ci: Fix pages workflow concurrency (#493)
Grouping on PR number or git ref (branch or tag). Without this all pages
builds were included in the same concurrency group, meaning that
concurrent
PR builds were leading to skipped workflows. As this is a required
check, that
meant that PRs could not be merged without opening/closing the PR - not
a
fun thing to have to do!