build: Fix image tags in release manifests (#516)
Previous releases incorrectly missed the `v` prefix in the image when
generating the `runtime-extensions-components.yaml`. This commit fixes
that always inculding the `v`.

This commit also ensures that the use of `trimPrefix` to remove the
`v` prefix is consistently applied. As documented at
https://goreleaser.com/customization/templates/#fn:version-prefix, the
`Version` variable does not include the `v` prefix, but this is only for
releases, and the `v` prefix is actually included in `Version` for
snapshot or nightly builds.