refactor: standardize the code for getting Helm values (#500)
I noticed there were 3 copies of `retrieveValuesTemplateConfigMap()` in
addition to the one generic one in `lifecycleutils`, refactored all of
these methods to reuse the existing function.

While refactoring I saw that for NFD the handler was adding a value,
moved the templating to the helm values. This copies the pattern in
other addons, but also makes it easier to override at cluster creation
time (kind of as a last resort if there is a bug in the handler code)

I also added `RetrieveValuesTemplate` to get the data instead of having
`valuesTemplateConfigMap.Data["values.yaml"]` in every handler.