refactor: provider an entrypoint to the infra provider meta handlers (#554)
**What problem does this PR solve?**:
This aligns the structure to how the lifecycle handlers are created.
More importantly it enables us to pass globalOptions to those infra
handlers that need it.

The goal of this is to use for the eventual refactor to support a more
generic kube-vip implementation. But for that I need to be able to read
a ConfigMap from the "default" namespace where all of these CMs are
created. Again similar to how its done for the CAAPH addons
https://github.com/nutanix-cloud-native/cluster-api-runtime-extensions-nutanix/blob/33b29dc368622777d8dbea004077532c2a84d27a/pkg/handlers/generic/lifecycle/nfd/strategy_helmaddon.go#L56-L61

**Which issue(s) this PR fixes**:
Refactor to support work for https://jira.nutanix.com/browse/D2IQ-100364

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->