refactor: Clean up API constants, and explain usage (#588)
**What problem does this PR solve?**:
We have not explained how we use API constants. We also have good
reasons to remove some constants. This PR does the former, and gets us
closer to the latter.

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->