build: Ensure provider metadata is up to date when releasing (#600)
When a new minor release is created, the metadata attached to the
release needs to include the release series provider compatibility
mapping. This commit ensures that if a new minor release is cut that the
`metadata.yaml` attached to the release is cotrect for that release.

I have manually fixed v0.8.1 and v0.7.0 releases and with this commit we
won't need to do it manually ever again!