fix: Correct the CSI handler logic (#603)
**What problem does this PR solve?**:
- Return an error when the CSI variable is defined, but the
  Providers list is empty.
- Return an error when the CSI provider is defined, but its
  StorageClassConfig list is empty.
- Return an error when any CSI provider is unknown.
- Return an error when DefaultStorage is empty.
- Return an error when the DefaultStorage Provider Name does not match a
set provider.
- Return an error when the DefaultStorage StorageClassConfig Name does
not match the name of a StorageClassConfig for the default provider.

Adds unit tests to verify this logic, and more. Covers the majority of
the generic CSI handler `AfterControlPlaneInitialized` hook.

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->
Adds a unit test; I think this is the first unit test in CAREN to
exercise a handler request-response flow.
 
**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->
I noticed these issues as I was studying the CSI handler in order to
write the ServiceLoadbalancer handler. I'll be adding a unit test to
#592 as well.

---------

Co-authored-by: Dimitri Koshkin <dimitri.koshkin@gmail.com>
Co-authored-by: Jimmi Dyson <jimmidyson@gmail.com>