refactor: Write configuration under /etc/caren
Previously, we wrote some configuration to /etc/cre. Now that the
project name is CAREN, we should use /etc/caren.

We also wrote containerd configuration to /etc/containerd/cre.d, but
this configuration is not read by containerd directly. Instead, it is
read by a script that merges it to the primary containerd configuration.
For that reason, this configuration belongs under /etc/caren.