refactor: Write configuration under /etc/caren (#656)
**What problem does this PR solve?**:
Previously, we wrote some configuration to /etc/cre. Now that the
project name is CAREN, we should use /etc/caren.

We also wrote containerd configuration to /etc/containerd/cre.d, but
this configuration is not read by containerd directly. Instead, it is
read by a script that merges it to the primary containerd configuration.
For that reason, this configuration belongs under /etc/caren.

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->