fix: Make Cluster the owner of image registry credential secret (#648)
**What problem does this PR solve?**:
Makes the Cluster resource the owner of image registry credential
Secret. CAREN creates this Secret for each Cluster, and it should be
deleted when the Cluster is deleted.

**Which issue(s) this PR fixes**:
Fixes https://jira.nutanix.com/browse/D2IQ-100572

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->
All test cases pass. The test is updated to create a Cluster resource,
and to provide the handler a client that can read a Cluster resource.

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->