fix: Post-process clusterconfig CRDs for supported CSI providers
Only supported CSI providers should be available on each CAPI provider. This
commit enforces that via the API, without changing handler logic to allow
for easier extensibility later, by simply updating the supported CSI provider
names in the API definitions.

This can only happen as CRD post-processing as kubebuilder annotations are
specified on API types in go. To use that approach for CSI providers here
would require changing the whole API structure to provide CAPI provider
specific types with CAPI provider specific annotations, which would be a
maintenance headache.

This issue arose for me when adding the `local-path` CSI provider for Docker
as I did not want this to be available on AWS or Nutanix clusters, as well
as the opposite way round for Docker clusters (no `aws-ebs` or `nutanix`
support).