refactor: rename credential provider response secret (#710)
**What problem does this PR solve?**:
Wanted to avoid "creds" abbreviation.
Also, the Konvoy secret names are
`<cluster-name>-image-registry-credentials` and
`<cluster-name>-image-registry-mirror-credentials` so it was pretty
confusing also having a `<cluster-name>-registry-creds`. Wanted to make
it a little more descriptive. (Suggestions welcome!)
Here is what that Secret looks like:
```
{
  "kind":"CredentialProviderResponse",
  "apiVersion":"credentialprovider.kubelet.k8s.io/v1",
  "cacheKeyType":"Image",
  "cacheDuration":"0s",
  "auth":{
    "docker.io": {"username": "<>", "password": "<>"}
  }
}
```

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->