fix: static credentials Secret generation (#717)
**What problem does this PR solve?**:
While working in this package I was confused why `createSecretIfNeeded`
was called that since it always generated a Secret.
Turns out there was a bug in these files that it never checked if the
passed in `configs []providerConfig` even have static credentials. Added
some failing unit tests and fixed the logic to skip generating the
Secret and the corresponding file if not needed.
While writing the tests, I also noticed that the templating was not
generating valid JSON for multiple registries. To keep that separator
logic simple I moved the `docker.io` alias into Go code.

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->
New unit tests and brought up a Nutanix cluster with `docker.io` creds.

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->