fix: Remove providerName from defaultStorage config (#696)
This commit removes the `providerName` from `defaultStorage`
configuration, just requiring that the `storageClassConfigName` exists
for one of the configured providers. This commit also adds validation to
ensure that the storage class configs have unique names across
providers, current behaviour could have led to conflicts if the storage
class name was used in multiple provider configurations.

When adding the `local-path` CSI provisioner for Docker clusters, I
found that the `providerName` in default storage class config was
redundant and led to unnecessary duplication of values. Removing this
unnecessary field reduces the duplication from:

```
defaultStorage:
  providerName: local-path
  storageClassConfigName: local-path
providers:
  - name: local-path
    storageClassConfig:
      - name: local-path
    strategy: HelmAddon
```

To:

```
defaultStorage:
  storageClassConfigName: local-path
providers:
  - name: local-path
    storageClassConfig:
      - name: local-path
    strategy: HelmAddon
```

While this may seem like only a small change, a single line removed, it
is always good to keep APIs as terse as possible while still retaining
strict behaviour. I think this commit achieves that.

Depends on #695.