refactor: Reusable HelmAddon strategy
This commit refactors some of the existing HelmAddon code used in various
handlers to a common applier, thus reducing code maintenance.

There are a number of other places where this could be applied, but they
currently have specific behaviour that is harder to generalise and as
such I have not refactored those handlers yet.