fix: adds a docker buildx step before release-snapshot (#741)
**What problem does this PR solve?**:

We add a make target to create a docker buildx builder container.

We create this builder container as a dependency of the `release` and `release-snapshot`
make targets, which fixes the local development workflow. This is tested on amd64 and arm64.

The new make target makes the two docker-related third-party github actions unnecessary, so we remove them. That brings the CI workflow one step closer to being reproducible locally.

---------

Co-authored-by: Daniel Lipovetsky <daniel.lipovetsky@nutanix.com>