fix: Namespace Sync controller should list no resources when source namespace is empty string (#725)
**What problem does this PR solve?**:
Previously, the controller listed resources in all namespaces when the
source namespace was an empty string. Now, it lists no resources.

The PR adds a test verifying the above. It also adds a test that
verifies that an the controller will not match any namespaces when the
target namespace label key is an empty string.

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->