test(e2e): Set empty env vars for Nutanix e2e vars (#749)
When running e2e tests for other providers, it is currently necessary
to export empty variables for `NUTANIX_ENDPOINT`, `NUTANIX_PASSWORD`
and `NUTANIX_USER`. This commit sets these to empty values in the
e2e configuration so they are not required for other providers e2e
tests but can still be overridden when running the Nutanix e2e tests.