fix: Namespacesync controller should reconcile an updated namespace (#775)
**What problem does this PR solve?**:
The namespacesync controller did not reconcile a namespace when it was
updated. We anticipated namespaces to meet the criteria for "target
namespace" when they were created. We also wanted to avoid reconciling
namespaces unnecessarily.

This PR makes the controller reconcile an updated namespace, if the
answer to question "is this a target namespace," has changed. It also
adds an integration test (which fails without the subsequent fix
commit).

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->