feat: Add waiter for object (#777)
**What problem does this PR solve?**:
Implements a wait for a check to pass against a typed object.

We'll use this in some lifecycle handers, e.g., in a future change for
deploying ServiceLoadBalancer configuration to the remote cluster.

(This is a copy of #762. I had to close that after #755 added required
checks that can't be run from PRs from public forks. )

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->