fix: Clean up MetalLB pod security standards labels
This commit removes the `pod-security.kubernetes.io/audit` and
`pod-security.kubernetes.io/warn` labels as they are redundant when
specifying the `pod-security.kubernetes.io/enforce` label to the same
level.

Also add the `pod-security.kubernetes.io/enforce-version` label to
always enforce the latest pod security polify version, even on upgrade.
This is fine because we are specifying the omst privileged pod security
standard, `privileged`, and as such should be safe to always enfore the
latest policy version.