fix: Ensure MetalLB lifecycle handler is independent of other handlers
Previously, the MetalLB handler created custom resources on the workload
cluster. These required a webhook to be ready. The webhook required one
or more Nodes to be Ready. That required CNI to be deployed.

When CAPI called the MetalLB handler before the CNI handler, this
condition would never be satisified. No handlers after the MetalLB would
be called.