fix: skip UUID annotation webhook for clusters with nil topology (#860)
**What problem does this PR solve?**:
We only want this webhook to run on Clusters that are created with
cluster topology. This also fixes an issue when an existing Cluster that
was created before deploying CAREN and this webhook is moved to the
bootstrap cluster.
Currently it fails the move with 
```
  2024-08-14 11:14:55 ERR     err="unable to pivot to the to-cluster: error setting Cluster.Spec.Paused=true: action failed after 10 attempts: error patching Cluster default/e2e-preprovisioned-upgrade-573225644: admission webhook "cluster-validator.caren.nutanix.com" denied the request: missing cluster UUID annotation caren.nutanix.com/cluster-uuid"
```

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->
I'm also now wondering if we want the validate to fail when the
annotation is missing. Its possible that an existing Cluster that was
created using another ClusterClass that does not use CAREN and then when
CAREN gets installed, it prevents those existing Clusters from being
updated because of a missing annotation that won't even be used.
**Should this webhook be smarter and only validate and default Clusters
that are referencing a CC that is using CAREN?**

---------

Co-authored-by: Jimmi Dyson <jimmidyson@gmail.com>