fix: Ensure ClusterAutoscaler can write status CM in workload cluster
Previously workload ClusterAutoscaler deployments were configured to
write the status configmap to the same namesapce as the workload cluster
is in on the management cluster. However, ClusterAutoscaler actually
writes this configmap and acquires a lock in the specified namespace in
the workload cluster. If that namespace does not exist then CA will fail
to run correctly.

This commit updates the target namespace to use the default
`kube-system` which always exists on all clusters. This is the CA
default and as such is completely safe to use.