fix: Properly set additionalTrustBundle for Nutanix CCM
The helm values template contained an error meaning that the
additionalTrustBundle setting would have been templated as part of the
Helm install, rather than the CAAPH addon deployment, meaning that
specifying an additional trust bundle would not work for Nutanix CCM
deployments.

This commit fixes that and adds tests that use the same template as the
CAAPH addon would do from the Helm chart to ensure that the template is
now correct.