feat: Extract CAAPH values templates to files (#896)
This simplifies the Helm templating directives by not requiring
inception-style escaping of templating braces, e.g. `{{ "{{" }}``
which are very hard to read and can introduce bugs.

This PR also removes the duplicate helm values files currently being
used
to generate the CRS configmaps, and instead references the helm values
that are in the charts directory, which ends up with a reduction in LOC
in
the project to maintain.

I feel this is a simpler way to achieve the same goals as #819 but
without
duplicating files and keeping all chart source files in the charts
directory.

Blocked by #895.