feat: Auto renewal of control plane certificates patch
This commit enables configuration of automatic certificate renewal via
the Kubeadm Control Plane provider as detailed in
https://cluster-api.sigs.k8s.io/tasks/certs/auto-rotate-certificates-in-kcp.html.

Although there is a reasonably sized refactoring (including renames) the
API types in the api module, the variable schema as opposed to CAPI
clusters via the external patches hook contains only the changes for
auto-renewal of certificates. I feel this is OK as the main clients of
the API are the Cluster authors rather than other projects using this
via code at this point at least.