fix: Introduce startup probe to CAREN pod to reduce races
Introduce a more frequent startup probe polling every second, which once
it succeeds will trigger readiness and liveness probes to run at their
default intervals. This reduces chances of races in clusterctl checking
for readiness while keeping the overhead on the kubelet for checking
readiness once the pod is up to a minimum (default probe period).