test: wait for COSI controller to be ready (#1014)
**What problem does this PR solve?**:
Missed this in
https://github.com/nutanix-cloud-native/cluster-api-runtime-extensions-nutanix/pull/1008
since I thought that all HCP are already checked.

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->