fix: use republished COSI controller image (#1022)
**What problem does this PR solve?**:
The `gcr.io/k8s-staging-sig-storage` repository is being shut down.
Temporarily moving the image until upstream moves it to a more permanent
location.

This pulls in the changes:
- https://github.com/mesosphere/charts/pull/1542
- https://github.com/mesosphere/charts/pull/1543

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->
The e2e tests will verify the controller is running.
I've also verified the image by following the examples in
https://github.com/nutanix-cloud-native/cosi-driver-nutanix

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->