fix: don't generate empty _default containerd mirror file (#1042)
**What problem does this PR solve?**:
Now that we're fixing how the `_default` file is being generated just
for mirrors, we should skip creating the file if only registries are
provided.

Depends on #1039.

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->