fix: update Cilium version to 1.16.7 (#1056)
**What problem does this PR solve?**:
This version fixes an upstream bug
https://github.com/cilium/cilium/pull/37086



**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->
I branched `release/v0.27.x` from the last release and this branch does
not include:
*
https://github.com/nutanix-cloud-native/cluster-api-runtime-extensions-nutanix/pull/1054
*
https://github.com/nutanix-cloud-native/cluster-api-runtime-extensions-nutanix/pull/1055