feat: properly support kube-vip upgrades (#1062)
**What problem does this PR solve?**:
See
https://docs.google.com/document/d/1RolIuLXf5qD5l2lya0ZtY-ZcGkh07ctiRiG8jjXMthU/edit?usp=sharing
with additional details.

We ran into an issue where upgrading CAREN causes an unwanted rollout of
all control plane Nodes.
This happens when CAREN is shipped with a new kube-vip version and the
difference in the static Pod spec changes the `file` contents in KCP and
forces new Nodes to be created.

We do not want this behavior and only want this rollout to happen at
controlled times.

In a recent change we introduced a mechanism of making backend changes
by using versioned handlers.
https://github.com/nutanix-cloud-native/cluster-api-runtime-extensions-nutanix/pull/1039/commits/7bc90941f0f848d5be1b05cad87b8ae26b782bd9

This change follows a similar pattern. A new handler is introduced that
will instead of reading the kube-vip template from a global ConfigMap
will read it directly from the KCPTemplate associated with the
ClusterClass.
We can then continue to change the kube-vip version (or any other spec
changes there) and downstream projects will just pick up whatever
changes through a new ClusterClass.

Another important goal with this approach was to not require consumers
of CAREN to maintain some reference to the "correct" kube-vip template.
We want to keep upgrades as simple as possible by only changing the
Kubernetes version, and let CAREN handle its own business logic.

**Alternative Considered Solutions**

1. Add a ConfigMap ref of the static Pod template to the API. This would
require the clients to keep track of and update another field on
upgrades. We want to keep the upgrade changes as simple as possible and
leave this kind of implementation detail to CAREN.
2. Add logic to only apply new changes when upgrading. This approach was
considered for a previous scenario, but we decided against it due to
complexity and instead introduced versioned ClusterClass and templates.

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->
Tested locally by creating a cluster with the new and old handlers.

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->
This PR introduces a "v3" version of the handlers with the modified
kube-vip handler, and preserves the previous behavior from v2.