build: Update k8s.io/kubelet dep
This is skipped by dependabot which ignores `k8s.io/*` in order for
dependencies in order to retain compatibility with `sigs.k8s.io/controller-runtime`,
but as `controller-runtime` does not have dependencies on `k8s.io/kubelet` then it
gets skipped on update. This commit brings it up to date.

I have attempted to fix up dependabot config to allow this but I could not find
a working configuration.