fix: use a new version for control-plane and worker mutation handlers (#1104)
**What problem does this PR solve?**:
When adding
https://github.com/nutanix-cloud-native/cluster-api-runtime-extensions-nutanix/pull/1097,
we effectively enabled this handler by default through API defaults. I
wrongly assumed that the worker and control-plane handlers were behind a
new version of the handlers registration. And because its not, this
caused a rollout of all Machines when CAREN is upgraded on the
management cluster.

This PR puts the worker and control-plane handlers behind a new version
of the hooks and does not include the new `noderegistration` handler in
the older version of the hooks.

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->
Tested locally by upgrading the handler version and observing that
Machines were NOT rolled out. We will need some strategy to test this in
an e2e test, or possibly think of a different approach for this, e.g.
running multiple versions of the controller instead of embedding it in a
single handler.

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->