refactor: Register types with scheme using pattern from Cluster API
This pattern removes the dependency on controller-runtime from the API
module. It is recommended by Cluster API.

https://cluster-api.sigs.k8s.io/developer/providers/getting-started/implement-api-types#registering-apis-in-the-scheme

Because the api module contains external types that do have a dependency
on controller-runtime, we have to make a similar change to these types
upstream before we can remove our dependency on controller-runtime.